Fixing accressors and using require_relative for including files
